Debug directory conflicts with the Visual Studio build system.
Debug directory name conflicts with the Visual Studio built .gitignore
file, so I have renamed it to Utilities to avoid confusion.
